refactor: remove bus/request module
