--- src/bus/request.rs
@@ -1,87 +1,0 @@
-// LNP Node: node running lightning network protocol and generalized lightning
-// channels.
-// Written in 2020 by
-//     Dr. Maxim Orlovsky <orlovsky@pandoracore.com>
-//
-// To the extent possible under law, the author(s) have dedicated all
-// copyright and related and neighboring rights to this software to
-// the public domain worldwide. This software is distributed without
-// any warranty.
-//
-// You should have received a copy of the MIT License
-// along with this software.
-// If not, see <https://opensource.org/licenses/MIT>.
-
-#![allow(clippy::clone_on_copy)]
-
-use std::collections::BTreeMap;
-use std::fmt::{self, Debug, Display, Formatter};
-use std::iter::FromIterator;
-
-use amplify::Wrapper;
-use bitcoin::secp256k1::rand::{thread_rng, RngCore};
-use internet2::addr::NodeAddr;
-use internet2::{CreateUnmarshaller, Unmarshaller};
-use lazy_static::lazy_static;
-use strict_encoding::{StrictDecode, StrictEncode};
-
-use crate::bus::msg::Msg;
-
-lazy_static! {
-    pub static ref UNMARSHALLER: Unmarshaller<Msg> = Msg::create_unmarshaller();
-}
-
-#[derive(Eq, PartialEq, Hash, Clone, Debug, Display, StrictEncode, StrictDecode)]
-#[display("request_id({0})")]
-pub struct RequestId(pub u64);
-
-impl RequestId {
-    pub fn rand() -> RequestId {
-        let mut id = [0u8; 8];
-        thread_rng().fill_bytes(&mut id);
-        RequestId(u64::from_be_bytes(id))
-    }
-}
-
-pub type RemotePeerMap<T> = BTreeMap<NodeAddr, T>;
-
-#[derive(Wrapper, Clone, PartialEq, Eq, Debug, From, StrictEncode, StrictDecode)]
-#[wrapper(IndexRange)]
-pub struct List<T>(Vec<T>)
-where
-    T: Clone + PartialEq + Eq + Debug + Display + StrictEncode + StrictDecode;
-
-#[cfg(feature = "serde")]
-impl<T> Display for List<T>
-where
-    T: Clone + PartialEq + Eq + Debug + Display + serde::Serialize + StrictEncode + StrictDecode,
-{
-    fn fmt(&self, f: &mut Formatter<'_>) -> fmt::Result {
-        f.write_str(&serde_yaml::to_string(self).expect("internal YAML serialization error"))
-    }
-}
-
-impl<T> FromIterator<T> for List<T>
-where
-    T: Clone + PartialEq + Eq + Debug + Display + serde::Serialize + StrictEncode + StrictDecode,
-{
-    fn from_iter<I: IntoIterator<Item = T>>(iter: I) -> Self {
-        Self::from_inner(iter.into_iter().collect())
-    }
-}
-
-#[cfg(feature = "serde")]
-impl<T> serde::Serialize for List<T>
-where
-    T: Clone + PartialEq + Eq + Debug + Display + serde::Serialize + StrictEncode + StrictDecode,
-{
-    fn serialize<S>(
-        &self,
-        serializer: S,
-    ) -> Result<<S as serde::Serializer>::Ok, <S as serde::Serializer>::Error>
-    where
-        S: serde::Serializer,
-    {
-        self.as_inner().serialize(serializer)
-    }
-}
